+++ frontend/lib/ocr/receipt-parser.ts
@@ -1,11 +1,6 @@
 import { OCRResult } from './types'
 import { IParser, IReceiptParser } from './interfaces'
-import { AtmReceiptParser } from './receipts/atm-parser'
-import { RaudhahSwalayanReceiptParser } from './receipts/raudhah-parser'
-import { ShoppingReceiptParser } from './receipts/shopping-parser'
-import { AciakMartReceiptParser } from './receipts/aciak-parser'
-import { CitraSwalayanReceiptParser } from './receipts/citra-parser'
-import { MinaSwalayanReceiptParser } from './receipts/mina-parser'
+import { OpenAIReceiptParser } from './openai-parser'
 
 export class ReceiptParser implements IParser {
   context: 'Receipt' = 'Receipt'
@@ -13,19 +8,14 @@ export class ReceiptParser implements IParser {
 
   constructor(receiptParsers?: IReceiptParser[]) {
     this.receiptParsers = receiptParsers ?? [
-      new AtmReceiptParser(),
-      new RaudhahSwalayanReceiptParser(),
-      new AciakMartReceiptParser(),
-      new CitraSwalayanReceiptParser(),
-      new MinaSwalayanReceiptParser(),
-      new ShoppingReceiptParser(),
+      new OpenAIReceiptParser(),
     ]
   }
 
-  parse(text: string, timezoneOffset?: string, filename?: string): OCRResult {
+  async parse(text: string, timezoneOffset?: string, filename?: string): Promise<OCRResult> {
     for (const receiptParser of this.receiptParsers) {
       if (receiptParser.identify(text)) {
-        return receiptParser.parse(text, timezoneOffset, filename)
+        return await receiptParser.parse(text, timezoneOffset, filename)
       }
     }